LU-7429 tests: inject lfsck failure properly
In sanity-lfsck test_23, we need to inject some failure stub to
simulate the case of dangling name entry. But the original mean
may misguide the LFSCK to regard it as bad name entry.

On the other hand, the test scripts injects some test delay that
may cause the LFSCK to run slowly, so need more time to wait the
LFSCK status changes.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I36c6c1bd8d07d2af6680e18b6c722e3bfa3a0062