LU-8139 ofd: Make ofd.max_brw to take number prefixes
ofd.max_brw control is written to take input in megabytes,
but it does not have _mb in the name making it confusing.

Change the code to accept prefixes like K and M, but also assume
that if no prefix was specified, it means M.

Update display output to show M after the number too, though
that loses precision since new scheme of things allows you
to set fractional megabytes too.

Change-Id: I2ac5331698fe70a5626c1dd77217b868578fa66c
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>