LU-7117 osp: control RPC to be sent when recovery
If remote server is in recovery, then we only allow recovery
related thread(s) on current server to send OUT RPC to such
remote server, others, such as normal RPC service threads,
will be blocked or failed. Otherwise, some stale information
may be returned from the remote server because of uncompleted
recovery on it, as to misguide subsequent operations.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Id9ac542751cc0042fba0a94166dfc57ace52dc69