LU-8171 hsm: stack overrun in hai_dump_data_field
With the stock 3.10 kernel, hai_dump_data_field will do a stack buffer
overrun when cat'ing /proc/fs/lustre/.../hsm/actions if an action has
some data in it.

hai_dump_data_field uses snprintf. But there is no check for
truncation, and the value returned by snprintf is used as-is.  The
coordinator code calls hai_dump_data_field with 12 bytes in the
buffer. The 6th byte of data is printed incompletely to make room for
the terminating NUL. However snprintf still returns 2, so when
hai_dump_data_field writes the final NUL, it does it outside the
reserved buffer, in the 13th byte of the buffer. This stack buffer
overrun hangs my VM.

This doesn't happen on CentOS 6.

Fix by checking that there is enough room for the next 2 characters
plus the NUL termonator. Don't print half bytes. Change the format to
02X instead of .2X, which makes more sense.

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: Icb6e658fc63f441b7a698f595ac71a1236ff3588