LU-8197 ptlrpc: do not reduce replay request deadline
New timeout for a replay request sent to a client with early reply is
calculated without taking into account that new deadline for the
replay request on server side is calculated based upon
obd_recovery_timeout. That results in reducing replay deadline which
forces to early disconnect of the client.

Calculate the timeout sent in early reply to a replay request in
according to new deadline of the replay request on server side.

Seagate-bug-id: MRP-3249
Signed-off-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Change-Id: I478ca84a9627816be886a077170285f8af4bfc29