LU-8194 test: Retry cancelling LRU locks.
The test sanity 127a tried to cancel lru locks for the read
operation to be performed. But in some cases the locks couldn't
be canceled because the locks was already being canceled by some
other thread and hence the count of canceled locks that is
returned is lesser than the actual number of unused locks. This
caused failure later as the test proceeds and the lock is not
canceled on time.

The fix includes a loop in the test which retries for 2 times
in the interval of 1 second. There is also optimization for when the
number of unused lock is already zero and also checks if it
becomes zero later. In case if the locks aren't canceled, the
error value which was returned was incorrect which is not
resolved.

Test-Parameters: testlist=sanity

Change-Id: I0c4dbea366b0062340d13afc7a3170c89e745d78
Seagate-bug-id: MRP-2735
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>