LU-8203 tests: Wait for archive operation to be completed.
As bandwidth control is set to 1048576B/s in the test, so
the archive operation takes <= 40 secs to copy 39000000
bytes.

If we don't wait for the ongoing archive action to finish
then status will remain as STARTED and never becomes
SUCCEED as we call copytool_cleanup soon after archiving
5MB of data and the request will be stuck there for ever.

Sometimes because of previous unclean shutdown of copytool,
test_60 results in failure because of stuck request.

Test-Parameter: trivial testlist=sanity-hsm,sanity-hsm

Change-Id: I2413c32a617e43fb24e7ca6696026a3c9e1cad00
Seagate-bug-id: MRP-2586
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>