LU-7334 lprocfs: Provide unitless numbers as a default unit
A few proc files expect a string with only numbers, but will still
accept units if provided. A new default unit '1' is accepted by lprocfs_str_with_units_to_s64 to indicate that a number without a
unit suffix is expected, but a unit is accepted if present at
the end of the string.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>
Change-Id: Iebf29f5d3c36e583729ea717288850729b99eeed