LU-7334 lprocfs: Few proc files are numbers, but use units.
A few proc files expect a string with only numbers, but will still
accept units if a user specifies them. Added a new default unit '1'
to lprocfs_str_with_units_to_s64 to indicate that a number without
unit suffix is expected, but a unit is still acceptable if present
at the end of the string being parsed.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>
Change-Id: Iebf29f5d3c36e583729ea717288850729b99eeed