LU-8220 tests: stripe file over all active OSTs
In test_iozone/parallel-scale-cifs, the iozone command creates the
temporary files under test with default stripe count=1, that is,
both the 256mb files reside on the same OST, over-utilizing the OST,
resulting in Use%=100%(No space left on device).

The fix is to create a sub-directory in the samba server mount
point with a stripe count of -1 (stripe over all available OSTs)
and then run iozone with this sub-directory path, so that all the
temporary files created within this sub-directory, during iozone,
reside on multiple OSTs and OST free space is utilized uniformly.

Since the test suite mounts lustre client on smb server, the "lfs
setstripe" command is executed on the smb server.

Test-Parameters: testlist=parallel-scale-cifs

Seagate-bug-id: MRP-2179
Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Change-Id: If225b62848b5266dc51ef791ef0d9c8002f908a5