LU-8399 osp: non-blocked osp_invalidate
The OSP transcation callback will purge the OSP object attribute
cache if related OUT RPC fail to update the object on remote MDT.
Under such case, it should use non-blocked version method to find
the object locally. Because if related object is not in cache or
the object to be purged from RAM, then needs to do nothing.

On the other hand, if the object itself to be purged from RAM and
if we use blocked method to locate such object for osp_invalidate,
then the OSP transaction callback will be blocked. If someone holds
the reference on such object and waiting for the OSP transaction
callback, then deadlock.

Be as improvement, we introduce new flag LOC_F_SEARCH_ONLY for
finding object that allow us to search the LU objects only without
allocating new objects in RAM if not found the target.

Test-Parameters: testlist=sanity-scrub,sanity-scrub,sanity-scrub,sanity-scrub
Test-Parameters: testlist=sanity-scrub,sanity-scrub,sanity-scrub,sanity-scrub
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I043182b349b38153cce4e20b2675556ead94c6c1