LU-8207 scripts: add auto-stripe option to lfs_migrate
Add a "-A" flag to lfs_migrate, which will automatically select the
stripe count as the file is rewritten. Initial algorithm to
determine stripe count is Log2(size_in_GB)+1, with an additional cap
on object size, though the algorithm or thresholds could conceivably
change in the future.  The primary intent for this feature is to be
able to give users a tool to fix stripe settings on existing files
based on file size.

Also includes a useful "-v" option to increase verbosity and help
test/debug/monitor what is being done.

Signed-off-by: Nathan Dauchy <nathan.dauchy@nasa.gov>
Change-Id: I9ce8b64e028d9abb66b6b49cf7675263fd7202f0