LU-8235 scripts: add pool selection option to lfs_migrate
Add a "-p" flag to lfs_migrate, which may be used to specify the
storage pool to which a file is migrated.  Performs sanity checks
on the number of targets in the pool, to ensure the desired stripe
count can be satisfied within the pool.

Also includes a useful "-v" option to increase verbosity and help
test/debug/monitor what is being done.

Signed-off-by: Nathan Dauchy <Nathan.Dauchy@noaa.gov>
Change-Id: Ia3f56bf528d2ac8155f08d93d01abd8bb9168cc4