LU-4752 test: Correct error_ignore message
The error_ignore routine requires a bug number to indicate why an
error should be ignored. In sanity test 248, there is no ticket
indicating why the test results should be ignored, but, due to running
the test in a VM, the performance numbers may not be what is
expected and, thus, the results can be ignored.

Add a reason to ignore the error to the call to error_ignore.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ieda9d06facd3e1519ac6a84e15505bc4c74aa60a