LU-6245 libcfs: remove types abstraction from libcfs/LNet code
Originally when lustre code was built for userland we needed
a proper way to handle 32 bit and 64 bit platforms when
reporting unsigned longs. Now that this code is only built
for kernel space and the kernel has it own special string
handling functions we don't need this abstraction anymore.
The libcfs types.h is not removed yet so not to break
any other patches ready to land. Had checkpatch.pl updated
to report to developers to not use the LP* abstraction.
This patch only updates the LNet/libcfs layers.

Change-Id: I0a338b4eb181d86b9c7d0f535d8c12c9f7045fb3
Signed-off-by: James Simmons <uja.ornl@yahoo.com>