LU-7225 llite: ladvise struct proposal
This is in response to discussion in LU-7225.  This initial
commit is just a view of the proposed protocol change, to
get feedback on that.  This will not even build.

Note I added api versions of the two structs, which is what
userspace will pass in, separate from what goes on the
network.

Pending feedback, I will produce a complete version of this
patch.

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: I7ac18e546f16a20c3c6bc6849becb0d45e3d5dc9