LU-7734 lnet: remove dependency on configuration changes
Remove dependency on peer_ni or local_ni changes in the selection
algorithm on configuration changes. After finding out the proper
CPT to use for sending a message, we unlock the currently held cpt
and relock the new one. By grabbing the refcount on both the
peer_ni and the local_ni before dropping the cpt, we ensure that
they don't get deleted from underneath us, and therefore we do not
need to re-run the algorithm

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Ie71eebc2301601cf1c85c6248dbed06951b89274