LU-8265 ldiskfs: fix one possible deadlock
This bug was originally motivated by porting project
quota support to ldiskfs:

ext4_setattr
 down_read(&EXT4_I(inode)->i_data_sem);
  ext4_mark_inode_dirty
   ext4_expand_extra_isize
    ext4_expand_extra_isize_ea
     ext4_xattr_block_set
      down_read(&EXT4_I(inode)->i_data_sem);--->deadlock here.
      ext4_new_meta_blocks

This problem happen when we trying to upgrade ldiskfs to
support project quota which will expand ext4 inode size.
I think this problem might also happen when we try to upgrade
Lustre from rhel6 to rhel7 in the future, so we better fix it.

Signed-off-by: Wang Shilong <wshilong@ddn.com>
Change-Id: Ie175e15cda197d12ed4e9258504f35bb924079f4