LU-8264 tests: setstripe striping parameters not inherited
When a user specifies any striping parameters in ‘lfs setstripe’, all
striping parameters come from the client and are not inherited from
the parent directory. There are several tests in the ost-pools script
that should detected this lack of inheritance from the parent
directory.

In ost-pools.sh, modify

1. test 6 to check the directory and file created by calling
‘lfs setstripe –c -1’ with not pool specified. Also, modify the pool
to contain only half the available OSTs so that using ‘-c -1’ with set
stripe is easily able to be checked.
2. test 20 to check that files and directories inherit the correct
pool information and that striping parameters to ‘lfs setstripe’ does
not alter this inheritance.

Some other minor clean up to the ost-pools script like removing
unnecessary calls to return after a call to ‘error’ and add a little
more detail to an error message.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I316c094b03ef03965f1d3ba1510f51c26350dc4f