LU-8299 clio: correct arithmetic in cl_index
cl_index uses a right shift for division, but takes an
signed argument (type loff_t).  If the argument is ever
negative, a right shift doesn't do division.

Make cl_index take pgoff_t as its argument instead.
This is an unsigned type of the same size (and it's what it
returns anyway).

Signed-off-by: Patrick Farrell <paf@cray.com>
Change-Id: Ifa6d0f8809e4c0818f32f0eec81709ae5af599e0