LU-8313 tests: Addition and removal of tests from SLOW list
On the basis of recent analysis there are some subtests
that qualify to be slow tests and can be added to the
SLOW list and some do not qualify to be SLOW and can be
removed from SLOW list.

A test is deemed as Slow or not if it run for five minutes
or more. A document is attached with ticket LU-8313 for
data gathered for all Slow tests.

Signed-off-by: Saurabh Tandan <saurabh.tandan@intel.com>
Change-Id: I7ab075650240d97886bdaf40227536269e094dec