LU-6245 client: remove types abstraction from client code
Originally when lustre code was built for both user land
and kernel space we needed a proper way to handle 32 bit
and 64 bit platforms when reporting unsigned longs. Now 
that this code is only built for kernel space and the 
kernel has it own special string handling functions we
don't need this abstraction anymore. Remove this 
abstraction from the client side code.

Change-Id: Ic0c55a413237bdf57d60031c12d5d9b62fa39cef
Signed-off-by: James Simmons <uja.ornl@yahoo.com>