LU-4185 llite: Revise create with no open optimization.
Currently ll_lookup_nd just returns a negative when
we are trying to create somethign with no open (read mkdir).
This is all fine most of the cases, except if the directory
where we are trying to do it is now writeable by us.
In that case vfs_create would return EPERM seeing as how
a negative dentry means the create cannot proceed.
But in reality if there is an existing name there that
we just did not have cached, a proper return is EEXIST that
could only happen if we did do the lookup.

So amend the optimization to only take place if the
directory is writeable by us, otherwise do the full lookup.

Change-Id: I5d5125f00c9a21d46a7c39331d06d53e6b717ca6
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>