LU-8592 dne: locate MDT-object when invalid its cache
When remote ibits lock is cancelled, related cached attributes
on the MDT should be invalid. Originally, for the convenience,
we make the remote ibits lock's callback data to reference the
object, then mdt_remote_object_lock() can easily find the object
and invalid its cache. But such way may cause object's reference
trouble. And such reference also prevents the object being freed.
In fact, it is unnecessary to reference the object, instead, we
can locate it via its fid that is the lock resouce name. Because
we only wants to invalid the cached object's attributes. If the
object is not cached, then need to do nothing. For such purpose,
we introduce the new lu_object_conf flag: LOC_F_CACHE_ONLY. That
allows us to search the cached object only, will NOT create one
if not found in cache.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ifd7adbce9a68da537f592c64117f4ecafb0a9ec4