LU-5092 nodemap: save nodemaps to targets for caching
Modify nodemap config storage to save config to targets as well as
MGSes. This allows targets to start with the last received nodemap
configuration even if the MGS is not available. The config is
replaced by the MGS' config the next time the MGS is available.

Test-Parameters: fortestonly envdefinitions=PTLDEBUG=-1,DEBUG_SIZE=300 testlist=replay-single
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I1c5221815618fe0265908bfd900ba55f44d1021b