LU-8386 zfs: Use ZFS_MAXPROPLEN instead of ZFS_MAXNAMELEN
ZFS_MAXNAMELEN was removed on zfs's master branch.  We
substitue ZFS_MAXPROPLEN in its place.  We only used this
macro in the get and set zfs property functions, so this
should be a reasonable exchange.

ZFS_MAXPROPLEN has also been around for a long time, so there
is no need to extra logic to make lustre lustre backwards
compatible with older ZFS releases.

Change-Id: I7292194f1a4e0311477954f7f83a3c353e785ac8
Test-Parameters: trivial
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>