LU-7734 lnet: fix routing selection
When running through the best_ni selection, the algorithm walks
through each net in the peer net list. If a peer net is local
followed by a peer net which is remote, the current algorithm would
select the best_ni for the local network, but still set routing to
true which sends the message out of the wrong local ni, and it
never reaches the destination.

The fix is to keep track when a best_ni on a local network is
preferred over a best_ni for a gateway and set routing off.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I3bec4809cc7478cce533591f4519f0c262be40f0