LU-8257 utils: fix mtab symlink logic
Existing logic to check if /etc/mtab is symlinked to /proc
fails if the symlink is ../proc/self/mounts.  This is in fact
the case in many newer linux distros, for example recent Fedora
versions.

This mod updates the logic in mtab_is_proc() to give the correct
answer for both old and new style symlinks.
It also cleans up the exact text on some error messages.

Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I49989f580c32e6baf820b88f6f9e6432aeded43d