LU-8330 wiretest: Add IOCTL checking to wiretest
Add IOCTL checking (to ensure they don't change) to wiretest, but
don't add to ptlrpc ASSERTS.  Doing this, showed that the
OBD_IOC_GET_MNTOPT doesn't work in user-space as mntopt_t isn't
defined in user-space and throws an error if the appropriate header is
included.  This type is an enum, so should be the same as a long.

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I6edf4c8bdd501a62cc7214ad14d3bb92822ffb99