LU-7650 lnet: handle mixed page size configurations.
The current LNet layer doesn't handle configurations
were two nodes are communicating that use different
page sizes. One assumption is that the default 1MB
bulk data packets are made of 256 4K pages which is
hard coded with LNET_MAX_IOV. This is not the case
for systems that use 64K page sizes. For the 64K
page case the LNet layer will be creating and sending
a 16MB packet for no reason. The second problem
discovered in the ko2iblnd driver is the handling of
RDMA fragements. The ko2iblnd driver expects the
number of fragments to send to the remote server
will be the same number locally which is not the
case. If a 64K page node needs to send data to
a remote 4K page system it needs to break the data
into 4K chunks to send. This patch resolves all
these issues. Also the logic has been versed in
the LNet core for the maximum payload size. Currently
setting a payload size larger than 1MB is not
supported due to the way LNet hardcored the 256
fragment size. Now the fragment count is determined
by the payload size. In theory large than 1MB LNet
packets are now possible.

Change-Id: I5aa4a464a0320fbd1841f9ad3add810e7b4f124a
Signed-off-by: James Simmons <uja.ornl@yahoo.com>