LU-8433 nrs: Maximizing throughput via rules with dependency
In order to ensure efficient utilization of I/O resource, we
propose a dependency rule strategy. The command of a dependency
rule is shown as follow:
        "start ruleB <matchCondition> deprule=ruleA
                lowerrate=$r1 upperrate=$r2"

Where 'deprule' represents the rule name of the dependent rule,
which means, 'ruleB' depends on 'ruleA' ; the key 'lowerrate'
indicates the lower bound of RPC rate limited value while the
key 'upperrate' indicates the upper bound of RPC rate limited
value.
The principle is that the real RPC rate limited value of a rule
is dynamically adjusted between the lowerrate and upperrate to
obtain more I/O bandwidth according to the spare I/O capacity that
its dependent rule does not make full use of.

Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: Ida2dd6e05a0094458ba28ea023b5fff4f0644e31