LU-8324 hsm: prioritize HSM requests
This patch separates the coordinator's implementation from the way
it processes requests stored in its llog.

A new scan_policy is defined: "priority_policy"
This allows to define which types of requests are to be processed
first by the coordinator.
For example it seems a good idea to have RESTORE requests processed
before ARCHIVE requests (unless the fs is full) as a RESTORE is most
certainly issued by a human client.

The working principles of the policy can be found in the comments of
"mdt_coordinator_priority_policy.c".

As allowed by the interface (introduced by this patch), the priority
policy adds two proc files under the hsm proc directory:
- csp_priority_mask:  NOOP ARCHIVE [RESTORE] [REMOVE] [CANCEL]
- csp_priority_ratio: 80%/20% (high/low)

(default values)

The priority_mask defines which requests have which level of
priority. Currently the policy defines two levels of priority:
high and low. The policy also defines the proportion of high and
low priority requests that the coordinator will process each time
it iterates over its llog.

Both proc files are writable. Example:
$> echo "+ARCHIVE -REMOVE" > csp_priority_mask
$> echo "75" > csp_priority_ratio # Defines the high value

Scan policies can be dynamically switched by writing the name of
the policy to use in the proc file "scan_policy" (which can be found
in the hsm proc directory).

Signed-off-by: Quentin Bouget <quentin.bouget.ocre@cea.fr>
Change-Id: I7d0645583b3db06d18b75a9e076b752ddbf883a3