LU-7650 o2iblnd: handle mixed page size configurations.
Currently it is not possible to send LNet traffic between
two nodes using infiniband hardware that have different
page sizes for the case when RDMA fragments are used.
When two nodes establish a connection the maximum number
of RDMA fragments possible are transmitted to ensure
proper handshaking between both nodes. The issue is that
on 64K page systems it will transmit how many pages it
can send to a 4K page system assuming the page on the
other end is the same size. In the case of a 1MB packet
size on a 4K system the number of fragments will be 256.
When the 64K page node receives that a packet of 256
fragments arrived it will assume that the total packet
size is 16MB and refuse the connection since it is larger
that the maximum message size. The solution is report
over the wire the maximum fragments in 4K units independent
of the native page size. Once that value is received
translate the number of 4K units into units of the native
page size.

Change-Id: I5aa4a464a0320fbd1841f9ad3add810e7b4f124a
Signed-off-by: James Simmons <uja.ornl@yahoo.com>